reward_server/pyproject.toml
[project]
name = "reward-server"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.7.0",
    "transformers>=4.57.0",
    "tokenizers>=0.21.1",
    "vllm>=0.16.0",
    "flask",
    "ray",
    "pillow",
    "lpips",
    "scikit-image",
]
